% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Cleans names}
\usage{
clean_names(names, sep = ".", replace.all = TRUE, is.valid = FALSE)
}
\arguments{
\item{names}{character vector to be modified. This is coerced to character if necessary.
First processed using base::make.name}

\item{sep}{The string separator. Default is `.`.}

\item{replace.all}{Logical. If TRUE (default) will replace multiple consecutive occurrences
of the separator with a single instance.}

\item{is.valid}{Not used. Logical. Test if the final character vector is syntactically valid.}
}
\value{
clean_name
}
\description{
Simple extension of the function [base::make.name()].
}
\details{
It will pass the input vector to it but allow the use of other separators
other than `.`
}
\examples{

some_col <- c(
   "I once",
   "got   some weird",
   "tables",
   "from.a.collaborator")

clean_names(some_col)
clean_names(some_col, replace.all=F, sep="_")
clean_names(some_col, sep="_")
clean_names(some_col, sep="+")

}
